Move fatal() to nyfe_fatal() and include it in lib.
Turns out calling something a lib means it gets used as a lib, so now
we have libnyfe that can call fatal() which required consumers to implement
a similar fatal() function in their code.

With this move a consumer is able to set a callback that is called whenever
nyfe_fatal() is called.

Very much not ideal design but this is safe and better than changing contracts.